requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=10.0.0
openpyxl>=3.1.0
pandas>=2.0.0
django-cors-headers>=4.0.0
openai>=1.0.0
psycopg2-binary>=2.9.0

